tasks/airport-lost-found-system/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
email-validator==2.2.0
redis==5.2.1
httpx==0.28.1
openai==1.59.7
azure-storage-blob==12.24.0
azure-search-documents==11.5.2
azure-ai-vision-imageanalysis==1.0.0
azure-communication-email==1.0.0
azure-communication-sms==1.0.1
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
azure-monitor-opentelemetry==1.6.4
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-instrumentation-httpx==0.50b0
qrcode==8.0
pillow==11.0.0
imagehash==4.3.1
